Tweak --archive-mirror semantics, repo printing
Unless --skip-download is specified "cache" is added to the list of
archive mirrors. This seems sensible to me as when you have local
archives downloaded you want people to use them. This way with
--skip-download and no --archive-mirror you can achieve an empty
archive-mirrors field.

The printing of the repo file has been reworked to use opam-file-format.
The printing mimics what opam-file (which depends on unix, sadly) does:
when archive-mirrors is an empty list the variable is elided; if it's a
singleton it's printed as a string; otherwise it's printed as a list.
This hopefully also ensures proper escaping.